// src/app/Models/company-interface.ts
export interface CompanyInterface {
  datasetid: string;
  recordid: string;
  fields: {
    efetcent: number;
    l6_declaree: string;
    l6_normalisee: string;
    libtefen: string;
    libcom: string;
    dapen: string;
    date_deb_etat_adm_et: string;
    dapet: string;
    ind_publipo: string;
    dcren: string;
    esaann: string;
    efencent: number;
    epci: string;
    esaapen: string;
    categorie: string;
    tcd: string;
    libtefet: string;
    tca: string;
    proden: string;
    libtu: string;
    siren: string;
    esasec4n: string;
    libtca: string;
    apet700: string;
    tu: string;
    typvoie: string;
    libvoie: string;
    defen: string;
    libapet: string;
    depcomen: string;
    esasec1n: string;
    siret: string;
    l4_normalisee: string;
    prodet: string;
    libactivnat: string;
    nomen_long: string;
    codpos: string;
    esasec3n: string;
    l1_declaree: string;
    lieuact: string;
    l4_declaree: string;
    amintren: string;
    origine: string;
    apen700: string;
    siege: string;
    nic: string;
    libreg_new: string;
    arronet: string;
    libmoden: string;
    numvoie: string;
    nicsiege: string;
    libmodet: string;
    libmonoact: string;
    ctonet: string;
    diffcom: string;
    dcret: string;
    du: string;
    coordonnees: [
      number,
      number
      ];
    depet: string;
    amintret: string;
    datemaj: string;
    modet: string;
    l7_normalisee: string;
    defet: string;
    activnat: string;
    uu: string;
    monoact: string;
    libnj: string;
    tefen: string;
    l1_normalisee: string;
    rpen: string;
    esasec2n: string;
    libapen: string;
    zemet: string;
    libesaapen: string;
    saisonat: string;
    nj: string;
    tefet: string;
    moden: string;
    comet: string;
    auxilt: string;
    l2_declaree: string;
    liborigine: string;
    rpet: string;
  };
  geometry: {
    type: string;
    coordinates: {
      0: number;
      1: number;
    }
  };
  record_timestamp: string;
}